leagacy code 삭제 및 폴더 구조 변경
