+++ assets/js/sign/config/errorMessage.js
@@ -11,6 +11,7 @@ const errorMessages = {
   PASSWORD_TOO_SHORT: '비밀번호는 8자 이상 20자 이하여야 합니다.',
 
   FAILED_SIGNUP: '시스템 장애로 회원가입에 실패하였습니다. 잠시 후 다시 시도해주세요.',
+  INVALID_NETWORK_ERROR: '시스템 장애로 로그인에 실패하였습니다. 잠시 후 다시 시도해주세요',
 };
 
 const getErrorMessage = errCode => errorMessages[errCode];
