+++ assets/js/sign/config/regex.js
