+++ assets/js/sign/view/loginView.js
@@ -1,4 +1,4 @@
-import getErrorMessage from '../../errorMessage.js';
+import getErrorMessage from '../config/errorMessage.js';
 
 export default class LoginView {
   constructor() {
