+++ signin.html
@@ -42,6 +42,6 @@
         </div>
       </div>
     </div>
-    <script type="module" src="/assets/js/sign/mvc/login.js"></script>
+    <script type="module" src="/assets/js/sign/login.js"></script>
   </body>
 </html>
