+++ signup.html
@@ -50,6 +50,6 @@
         </div>
       </div>
     </div>
-    <script type="module" src="/assets/js/sign/mvc/signUp.js"></script>
+    <script type="module" src="/assets/js/sign/signUp.js"></script>
   </body>
 </html>
